+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-6215] Use td-core-lib for elastic search and reindex
+
 ## [5.19.0] 2023-11-28
 
 ### Changed